add pacer based on wall clock instead of relative differences in duration
The reason for the change is that errors in pacing add up if we pace relative to
the previous request, but do not add up if we always do our calculations relative
to the start of the phase.

Like measuring 1 meter by measuing 2 millimeters on top of each other 500 times
will not be as accurate as measuring the 1 meter in one go.

The source of errors is simply execution time.